+++ src/data/id/handle/generation.rs
@@ -122,7 +122,7 @@ macro_rules! handle_gen· {
             const fn __allowed<P, I>()
             where
                 P: $crate::PrimIndex,
-                I: $crate::IndexRepr<Prim = P>,
+                I: $crate::ReprIndex<Prim = P>,
             {}
             __allowed::<$P, $I>();
         };