+++ src/num/grain/_.rs
@@ -1,34 +1,32 @@
 // devela/src/num/grain/_.rs
 //
 #![doc = crate::_DOC_NUM_GRAIN!()] // public
-#![doc = crate::_doc!(modules: crate::num; grain: niche, wide)]
+#![doc = crate::_doc!(modules: crate::num; grain: niche, prim, wide)]
 #![doc = crate::_doc!(flat:"num")]
 #![doc = crate::_doc!(extends: num)]
 //
 
 crate::mods_in! {
         // mod_ big; // TODO
-        mod_ cast; // Cast, PrimCast, PrimJoin, PrimSplit, cast!
         // mod_ float; // Compact binary floating-point formats and representations
         mod_ lim; // Boundary-aware integer representations and arithmetic
-    pub mod_ niche; // MaybeNiche, NonNiche*, NonZero*, Non<Max|Min|Value>*, nm!, nv!, nz!
-        mod prim; // PrimFloat, PrimInt, PrimScalar, PrimSint, PrimUint
-    pub mod_ wide; // lane!
+    pub mod_ niche; // Specialized numeric types and behaviors
+    pub mod_ prim; // Primitive numeric types, families, and representations
+    pub mod_ wide; // Wide numeric types and parallel arithmetic
 }
 crate::mods_out! { // _mods, _pub_mods, _crate_internals, _hidden
     _mods {
         #[doc(inline)]
         pub use super::{
             // big::_all::*,
-            cast::_all::*,
-            // float::_all::*, // RELOCATE
+            // float::_all::*,
             lim::_all::*,
-            prim::*,
         };
     }
     _pub_mods {
         pub use super::{
             niche::_all::*,
+            prim::_all::*,
             wide::_all::*,
         };
     }