+++ src/num/grain/prim/cast/namespace/join.rs
@@ -1,4 +1,4 @@
-// devela/src/num/grain/cast/namespace/join.rs
+// devela/src/num/grain/prim/cast/namespace/join.rs
 //
 //! Construct an unsigned primitive by joining an array of smaller unsigned primitives.
 //