+++ src/num/grain/prim/cast/namespace/split.rs
@@ -1,4 +1,4 @@
-// devela/src/num/grain/cast/namespace/split.rs
+// devela/src/num/grain/prim/cast/namespace/split.rs
 //
 //! Functions to split a primitive into an array of smaller primitives.
 //