+++ src/num/grain/prim/cast/traits/_.rs
@@ -1,4 +1,4 @@
-// devela/src/num/grain/cast/traits/_.rs
+// devela/src/num/grain/prim/cast/traits/_.rs
 //
 //! Helpers for converting between primitives.
 //