+++ src/num/grain/prim/cast/traits/join.rs
@@ -1,4 +1,4 @@
-// devela/src/num/grain/cast/traits/join.rs
+// devela/src/num/grain/prim/cast/traits/join.rs
 //
 //! construct an unsigned primitive by joining an array of smaller unsigned primitives.
 //
@@ -11,7 +11,7 @@ use crate::Cast;
 #[doc = crate::_tags!(num primitive)]
 /// Offers methods to construct a primitive from an array or slice of smaller primitives.
 #[doc = crate::_doc_meta!{
-    location("num/grain", trait PrimJoin),
+    location("num/grain/prim", trait PrimJoin),
 }]
 /// Methods expecting an array are more efficient than the ones expecting an
 /// slice. On the other hand slices of any lenght are supported as follows: