+++ src/num/grain/prim/cast/traits/split.rs
@@ -1,4 +1,4 @@
-// devela/src/num/grain/cast/traits/split.rs
+// devela/src/num/grain/prim/cast/traits/split.rs
 //
 //! fns to split a primitive into an array of smaller primitives.
 //
@@ -8,7 +8,7 @@ use crate::Cast;
 #[doc = crate::_tags!(num primitive)]
 /// Offers methods to split a primitive into an array of smaller primitives.
 #[doc = crate::_doc_meta!{
-    location("num/grain", trait PrimSplit),
+    location("num/grain/prim", trait PrimSplit),
 }]
 /// See also the [`Cast`] type for the equivalent *const* methods, and the
 /// [`PrimJoin`][super::PrimJoin] trait for the opposite operations.